requirements.txt
# requirements.txt
fastapi
uvicorn
authlib[asyncio]
httpx
python-dotenv